Modify GResponseCache implementation to guarantee unique cache keys
The GResponseCache now implements a three-level response cache, with the first level being the name, the second being a hash value of the true and measured energies, and the third level a hash value of the instrument direction. The hash values are implemented as uint64_t 64 Bit unsigned integers.

+++ include/GInstDir.hpp
@@ -29,6 +29,7 @@
 
 /* __ Includes ___________________________________________________________ */
 #include <string>
+#include <cstdint>
 #include "GBase.hpp"
 
 
@@ -59,7 +60,7 @@ public:
     virtual void        clear(void) = 0;
     virtual GInstDir*   clone(void) const = 0;
     virtual std::string classname(void) const = 0;
-    virtual double      hash(void) const = 0;
+    virtual uint64_t    hash(void) const = 0;
     virtual std::string print(const GChatter& chatter = NORMAL) const = 0;
 
 protected:
